Rename jpa package to "entities" package Change mapping for findAllVacanciesWithFilter to "/" Granted access to non logged in users to vacancy search page
