+++ src/main/java/world/inetum/realdolmen/inetumrealJobs/repositories/RoleRepository.java
@@ -5,8 +5,8 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import world.inetum.realdolmen.inetumrealJobs.jpa.ERole;
-import world.inetum.realdolmen.inetumrealJobs.jpa.Role;
+import world.inetum.realdolmen.inetumrealJobs.entities.ERole;
+import world.inetum.realdolmen.inetumrealJobs.entities.Role;
 
 
 @Repository
