+++ src/main/java/world/inetum/realdolmen/inetumrealJobs/repositories/UserRepository.java
@@ -3,7 +3,7 @@ package world.inetum.realdolmen.inetumrealJobs.repositories;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import world.inetum.realdolmen.inetumrealJobs.jpa.User;
+import world.inetum.realdolmen.inetumrealJobs.entities.User;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
