+++ src/main/java/world/inetum/realdolmen/inetumrealJobs/security/WebSecurityConfig.java
@@ -56,7 +56,9 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
         http.cors().and().csrf().disable()
                 .exceptionHandling().authenticationEntryPoint(unauthorizedHandler).and()
                 .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS).and()
-                .authorizeRequests().antMatchers("/authentication/**").permitAll()
+                .authorizeRequests()
+                .antMatchers("/authentication/**").permitAll()
+                .antMatchers("/api/vacancies/**").permitAll()
                 .antMatchers("/api/test/**").permitAll()
                 .anyRequest().authenticated();
