+++ src/main/java/world/inetum/realdolmen/inetumrealJobs/services/UserDetailsImpl.java
@@ -10,7 +10,7 @@ import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import world.inetum.realdolmen.inetumrealJobs.jpa.User;
+import world.inetum.realdolmen.inetumrealJobs.entities.User;
 
 public class UserDetailsImpl implements UserDetails {
     private static final long serialVersionUID = 1L;
